+++ config/environments/production.rb
@@ -59,7 +59,10 @@ Rails.application.configure do
 
   # Use the lowest log level to ensure availability of diagnostic information
   # when problems arise.
-  config.log_level = :debug
+  config.log_level = :info
+
+  # Use lograge on production to minimise logging
+  config.lograge.enabled = true
 
   # Prepend all log lines with the following tags.
   config.log_tags = [:request_id]