+++ CHANGELOG.md
@@ -175,6 +175,7 @@
 - XML download contains a `narrative` element for region & country containing the region or country name
 - Ingest AMS GCRF data from IATI
 - When a programme has an extending organisation set, the same organisation is set as the implementing organisation 
+- Use the `lograge` gem to reduce logging in production
 
 [unreleased]: https://github.com/UKGovernmentBEIS/beis-report-official-development-assistance/compare/release-8...HEAD
 [release-8]: https://github.com/UKGovernmentBEIS/beis-report-official-development-assistance/compare/release-7...release-8